Fix CreateMeal to mealtype=5

+++ NutriDbService/Helpers/MealHelper.cs
@@ -63,11 +63,12 @@ namespace NutriDbService.Helpers
                     MealTime = DateTime.UtcNow.ToLocalTime().AddHours(3)//IsTyme ? parseTime : DateTime.UtcNow.ToLocalTime().AddHours(3)
                 };
 
-                var oldmeal = _nutriDbContext.Meals.SingleOrDefault(x => x.UserId == meal.UserId && x.MealTime.Date == meal.MealTime.Date && x.Type == meal.Type && x.Type != 5);
+                var oldmeal = _nutriDbContext.Meals.SingleOrDefault(x => x.UserId == meal.UserId && x.MealTime.Date == meal.MealTime.Date && x.Type == meal.Type);
                 if (oldmeal != null)
                 {
                     _nutriDbContext.Database.BeginTransaction();
-                    _nutriDbContext.Dishes.RemoveRange(_nutriDbContext.Dishes.Where(x => x.MealId == oldmeal.Id));
+                    if (meal.Type != 5)
+                        _nutriDbContext.Dishes.RemoveRange(_nutriDbContext.Dishes.Where(x => x.MealId == oldmeal.Id));
                     foreach (var di in dishes)
                         di.MealId = oldmeal.Id;
